add main, fix leak in parse

+++ srcs/setting/set_mlx.c
@@ -14,14 +14,14 @@
 
 void		set_mlx(
 	t_dispatcher *dispatcher,
-	float width,
-	float height
+	int width,
+	int height
 )
 {
 	dispatcher->p_mlx = mlx_init();
 	dispatcher->p_win = mlx_new_window(
-		dispatcher->p_mlx, (int)width, (int)height, "rt");
+		dispatcher->p_mlx, width, height, "rt");
 	mlx_key_hook(dispatcher->p_win, &key_press, dispatcher);
 	dispatcher->p_img = mlx_new_image(
-		dispatcher->p_mlx, (int)width, (int)height);
+		dispatcher->p_mlx, width, height);
 }
